requirements.txt
fastapi
uvicorn
dynaconf
alembic
bcrypt
passlib
PyMySQL
python-jose
python-multipart
SQLAlchemy
sqlacodegen